+++ nifi-nar-bundles/nifi-couchbase-bundle/nifi-couchbase-services-api/pom.xml
@@ -22,25 +22,18 @@
         <version>1.5.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>nifi-couchbase-nar</artifactId>
-    <version>1.5.0-SNAPSHOT</version>
-    <packaging>nar</packaging>
-    <properties>
-        <maven.javadoc.skip>true</maven.javadoc.skip>
-        <source.skip>true</source.skip>
-    </properties>
+    <artifactId>nifi-couchbase-services-api</artifactId>
+    <packaging>jar</packaging>
 
     <dependencies>
         <dependency>
             <groupId>org.apache.nifi</groupId>
-            <artifactId>nifi-couchbase-services-api-nar</artifactId>
-            <type>nar</type>
+            <artifactId>nifi-api</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.apache.nifi</groupId>
-            <artifactId>nifi-couchbase-processors</artifactId>
-            <version>1.5.0-SNAPSHOT</version>
+            <groupId>com.couchbase.client</groupId>
+            <artifactId>java-client</artifactId>
+            <version>2.2.0</version>
         </dependency>
     </dependencies>
-
 </project>
